# perf: use `Get-CimInstance` to guess editor on windows