--- a/packages/launch-editor/guess.js
+++ b/packages/launch-editor/guess.js
@@ -37,17 +37,15 @@ module.exports = function guessEditor (specifiedEditor) {
       }
     } else if (process.platform === 'win32') {
       const output = childProcess
-        .execSync('powershell -Command "Get-Process | Select-Object Path"', {
-          stdio: ['pipe', 'pipe', 'ignore']
-        })
+        .execSync(
+          'wmic process where "executablepath is not null" get executablepath',
+          {
+            stdio: ['pipe', 'pipe', 'ignore']
+          }
+        )
         .toString()
       const runningProcesses = output.split('\r\n')
       for (let i = 0; i < runningProcesses.length; i++) {
-        // `Get-Process` sometimes returns empty lines
-        if (!runningProcesses[i]) {
-          continue
-        }
-
         const fullProcessPath = runningProcesses[i].trim()
         const shortProcessName = path.basename(fullProcessPath)