@sapphi-red 2022-06-17
It seems `wmic` could be removed and it is removed on Windows 11.
I will change this PR to use powershell.

Also `Get-WmiObject` was removed from Powershell 6 (Windows 10 includes powershell 5 so it works) so I will use `Get-CmiInstance` instead (this could be used from Powershell 3 (2012/09)).