@nrayburn-tech 2022-08-01
Tested this on my computer and had similar perf improvements.

Not sure if it is worth investigating since it is native and might add some complexity, but vscode uses https://github.com/microsoft/vscode-windows-process-tree as a replacement to `wmic` and it claims to have better performance than `wmic`.