@sodatea 2022-08-01
> Not sure if it is worth investigating since it is native and might add some complexity, but vscode uses [microsoft/vscode-windows-process-tree](https://github.com/microsoft/vscode-windows-process-tree) as a replacement to `wmic` and it claims to have better performance than `wmic`.

Thanks for the information!
I think it's too heavy (14.7 MB unpacked) to be included in this package for now.
But it's good to know there's a backup package available if we encounter problems with the current approach one day.